resource-scripts/set_azure_mgmt_firewall/requirements.txt
cloudshell-automation-api
cloudshell-logging
cloudshell-sandbox-reporter
